--- a/apisix/cli/ngx_tpl.lua
+++ b/apisix/cli/ngx_tpl.lua
@@ -354,9 +354,6 @@ http {
     include mime.types;
     charset {* http.charset *};
 
-    # error_page
-    error_page 500 @50x.html;
-
     {% if http.real_ip_header then %}
     real_ip_header {* http.real_ip_header *};
     {% end %}
@@ -467,13 +464,6 @@ http {
                 apisix.http_control()
             }
         }
-
-        location @50x.html {
-            set $from_error_page 'true';
-            content_by_lua_block {
-                require("apisix.error_handling").handle_500()
-            }
-        }
     }
     {% end %}
 
@@ -547,13 +537,6 @@ http {
 
             content_by_lua_block {
                 apisix.http_admin()
-            }
-        }
-
-        location @50x.html {
-            set $from_error_page 'true';
-            content_by_lua_block {
-                require("apisix.error_handling").handle_500()
             }
         }
     }
@@ -655,7 +638,6 @@ http {
             set $upstream_host               $http_host;
             set $upstream_uri                '';
             set $ctx_ref                     '';
-            set $from_error_page             '';
 
             {% if wasm then %}
             set $wasm_process_req_body       '';
@@ -828,20 +810,6 @@ http {
             proxy_pass $upstream_mirror_uri;
         }
         {% end %}
-
-        location @50x.html {
-            set $from_error_page 'true';
-            content_by_lua_block {
-                require("apisix.error_handling").handle_500()
-            }
-            header_filter_by_lua_block {
-                apisix.http_header_filter_phase()
-            }
-
-            log_by_lua_block {
-                apisix.http_log_phase()
-            }
-        }
     }
     {% end %}