--- a/t/node/upstream-status-5xx.t
+++ b/t/node/upstream-status-5xx.t
@@ -206,8 +206,8 @@ passed
 --- request
 GET /server_error
 --- error_code: 500
---- response_body_like
-.*apisix.apache.org.*
+--- response_body eval
+qr/500 Internal Server Error/
 --- response_headers
 X-APISIX-Upstream-Status: 500