--- a/t/node/upstream-status-all.t
+++ b/t/node/upstream-status-all.t
@@ -209,8 +209,8 @@ apisix:
 --- request
 GET /server_error
 --- error_code: 500
---- response_body_like
-.*apisix.apache.org.*
+--- response_body eval
+qr/500 Internal Server Error/
 --- response_headers
 X-APISIX-Upstream-Status: 500