Performance info:

wrk2 -U -d 30 --timeout 10s -c 100 -t 5 -R 5000 http://127.0.0.1:9080/hello

Before:
Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.15ms    9.57ms  86.21ms   94.70%
    Req/Sec     1.00k     7.98     1.04k    91.21%
After:
Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.39ms    2.20ms  44.13ms   98.86%
    Req/Sec     1.05k   186.67     2.22k    79.22%

wrk -d 30 --timeout 10s -c 100 -t 5 http://127.0.0.1:9080/hello

Before:
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    16.29ms   10.12ms 343.24ms   99.07%
    Req/Sec     1.28k   133.95     2.63k    90.79%

After:
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    17.81ms   26.67ms 711.64ms   98.38%
    Req/Sec     1.29k   179.37     2.32k    90.25%

Signed-off-by: spacewander <spacewanderlzx@gmail.com>

### Description

<!-- Please include a summary of the change and which issue is fixed. -->
<!-- Please also include relevant motivation and context. -->

Fixes # (issue)

### Checklist

- [ ] I have explained the need for this PR and the problem it solves
- [x] I have explained the changes or the new features added to this PR
- [ ] I have added tests corresponding to this change
- [ ] I have updated the documentation to reflect this change
- [ ] I have verified that this change is backward compatible (If not, please discuss on the [APISIX mailing list](https://github.com/apache/apisix/tree/master#community) first)

<!--

Note

1. Mark the PR as draft until it's ready to be reviewed.
2. Always add/update tests for any changes unless you have a good reason.
3. Always update the documentation to reflect the changes made in the PR.
4. Make a new commit to resolve conversations instead of `push -f`.
5. To resolve merge conflicts, merge master instead of rebasing.
6. Use "request review" to notify the reviewer after making changes.
7. Only a reviewer can mark a conversation as resolved.

-->